LayoutTests/cssom/observable-array-setter-expected.txt

Tests for ObservableArray's [[Put]]

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS document.adoptedStyleSheets.length is 1
PASS descriptor.configurable is false
PASS descriptor.enumerable is false
PASS descriptor.writable is true
PASS descriptor.value is 1
PASS document.adoptedStyleSheets[0] is sheet1
document.adoptedStyleSheets.length = 0
PASS document.adoptedStyleSheets.length is 0
PASS document.adoptedStyleSheets[0] is undefined.
PASS document.adoptedStyleSheets.length is 2
document.adoptedStyleSheets.length = 1
PASS document.adoptedStyleSheets.length is 1
PASS document.adoptedStyleSheets[0] is sheet1
PASS document.adoptedStyleSheets[1] is undefined.
document.adoptedStyleSheets.length = 5
PASS document.adoptedStyleSheets.length is 1
PASS document.adoptedStyleSheets[0] is sheet1
PASS document.adoptedStyleSheets[1] is undefined.
PASS document.adoptedStyleSheets.length = 0.5 threw exception RangeError: Invalid length.
document.adoptedStyleSheets.foo = 1;
PASS document.adoptedStyleSheets.foo is 1
PASS successfullyParsed is true

TEST COMPLETE
